social_media_api/requirements.txt
Package                Version
---------------------- ----------------
appdirs                1.4.4
apturl                 0.5.2
asgiref                3.8.1
attrs                  21.2.0
bcrypt                 3.2.0
beautifulsoup4         4.10.0
blinker                1.4
Brlapi                 0.8.3
cattrs                 1.10.0
certifi                2020.6.20
chardet                4.0.0
click                  8.0.3
colorama               0.4.4
command-not-found      0.3
cryptography           3.4.8
cupshelpers            1.0
dbus-python            1.2.18
defer                  1.0.6
distlib                0.3.8
distro                 1.7.0
distro-info            1.1+ubuntu0.1
Django                 5.1
django_csp             3.8
django-filter          24.3
django-taggit          6.0.0
djangorestframework    3.15.2
duplicity              0.8.21
fasteners              0.14.1
filelock               3.15.4
future                 0.18.2
gunicorn               23.0.0
html5lib               1.1
httplib2               0.20.2
idna                   3.3
importlib-metadata     4.6.4
jeepney                0.7.1
keyman-config          17.0.328
keyring                23.5.0
language-selector      0.1
launchpadlib           1.10.16
lazr.restfulclient     0.14.4
lazr.uri               1.0.6
lockfile               0.12.2
louis                  3.20.0
lxml                   4.8.0
macaroonbakery         1.3.1
Mako                   1.1.3
MarkupSafe             2.0.1
monotonic              1.6
more-itertools         8.10.0
mysql-connector-python 9.0.0
mysqlclient            2.2.4
netifaces              0.11.0
numpy                  1.21.5
oauthlib               3.2.0
olefile                0.46
onboard                1.4.1
packaging              21.3
paramiko               2.9.3
pexpect                4.8.0
Pillow                 9.0.1
pip                    22.0.2
platformdirs           4.2.2
protobuf               3.12.4
psycopg2               2.9.9
ptyprocess             0.7.0
pycairo                1.20.1
pycups                 2.0.1
PyGObject              3.42.1
PyJWT                  2.3.0
pymacaroons            0.13.0
PyNaCl                 1.5.0
pyparsing              2.4.7
pyRFC3339              1.1
python-apt             2.4.0+ubuntu2
python-dateutil        2.8.1
python-debian          0.1.43+ubuntu1.1
python-magic           0.4.24
pytz                   2022.1
pyxdg                  0.27
PyYAML                 5.4.1
qrcode                 7.3.1
reportlab              3.6.8
requests               2.25.1
requests-cache         0.9.1
SecretStorage          3.3.1
sentry-sdk             1.4.3
setuptools             59.6.0
six                    1.16.0
soupsieve              2.3.1
sqlparse               0.5.1
systemd-python         234
typing_extensions      4.12.2
ubuntu-advantage-tools 8001
ubuntu-drivers-common  0.0.0
ufw                    0.36.1
unattended-upgrades    0.1
url-normalize          1.4.3
urllib3                1.26.5
usb-creator            0.3.7
vboxapi                1.0
virtualenv             20.26.3
wadllib                1.3.6
webencodings           0.5.1
wheel                  0.37.1
whitenoise             6.7.0
xdg                    5
xkit                   0.0.0
zipp                   1.0.0
